requirements.txt
alembic==1.4.2
asyncio==3.4.3
beautifulsoup4==4.8.0
blinker==1.4
bokeh==2.0.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Email==1.4.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
multitasking==0.0.9
numpy==1.17.2
packaging==20.3
pandas==0.25.1
Pillow==7.1.2
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
tornado==6.0.4
typing-extensions==3.7.4.2
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
yfinance==0.1.54
